upgraded to 8.3.6

+++ core/lib/Drupal/Core/Config/ConfigImporter.php
@@ -862,6 +862,7 @@ class ConfigImporter {
             $entity_storage = $this->configManager->getEntityManager()->getStorage($entity_type_id);
             $entity_type = $this->configManager->getEntityManager()->getDefinition($entity_type_id);
             $entity = $entity_storage->load($entity_storage->getIDFromConfigName($name, $entity_type->getConfigPrefix()));
+            if($entity)
             $entity->delete();
             $this->logError($this->t('Deleted and replaced configuration entity "@name"', ['@name' => $name]));
           }
